+++ driver/src/main/java/oracle/nosql/driver/ops/PutResult.java
@@ -110,8 +110,8 @@ public class PutResult extends WriteResult {
      *
      * @since 5.4.18
      */
-    public String getExistingRowMetadata() {
-        return super.getExistingRowMetadataInternal();
+    public String getExistingLastWriteMetadata() {
+        return super.getExistingLastWriteMetadataInternal();
     }
 
     /* from Result */